Add isTransparent parameter Add WIndowButtonBuilder
