+++ lib/src/widgets/widgets.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: directives_ordering
+
 // All the widgets imported from libadwaita
 export 'adw/action_row.dart';
 export 'adw/avatar.dart';
@@ -6,21 +8,20 @@ export 'adw/clamp.dart';
 export 'adw/combo_row.dart';
 export 'adw/flap.dart';
 export 'adw/header_bar.dart';
-export 'adw/new/header_button.dart';
-export 'adw/new/scaffold.dart';
-export 'adw/new/sidebar.dart';
-export 'adw/new/text_button.dart';
-// New Widgets that are not present in libadwaita
-export 'adw/new/text_field.dart';
 export 'adw/preferences_group.dart';
 export 'adw/view_stack.dart';
 export 'adw/view_switcher.dart';
 export 'adw/view_switcher_tab.dart';
+
+// New Widgets that are not present in libadwaita
+export 'adw/new/header_button.dart';
+export 'adw/new/scaffold.dart';
+export 'adw/new/sidebar.dart';
+export 'adw/new/text_button.dart';
+export 'adw/new/text_field.dart';
+export 'adw/new/window_button_builder.dart';
+
 // All the widgets imported from Gtk4
 export 'gtk/popup_menu.dart';
 export 'gtk/stack_sidebar.dart';
 export 'gtk/toggle_button.dart';
-// Window Buttons
-export 'window_buttons/close.dart';
-export 'window_buttons/maximize.dart';
-export 'window_buttons/minimize.dart';
